requirements.txt
# Core PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0
pdf2image>=1.16.3
fitz>=0.0.1.dev2  # PyMuPDF

# OCR and Image Processing
pytesseract>=0.3.10
opencv-python>=4.8.0
Pillow>=10.0.0

# Text Processing and NLP
spacy>=3.6.0
nltk>=3.8.1
regex>=2023.6.3
python-Levenshtein>=0.21.1

# Machine Learning and Clustering
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
hdbscan>=0.8.29
datasketch>=1.6.4
umap-learn>=0.5.3

# Database and ORM
psycopg2-binary>=2.9.6
sqlalchemy>=2.0.0
alembic>=1.11.0

# Web Framework and API
fastapi>=0.100.0
uvicorn>=0.23.0
jinja2>=3.1.2
python-multipart>=0.0.6
aiofiles>=23.1.0

# CLI and Configuration
typer>=0.9.0
click>=8.1.0
rich>=13.4.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# HTTP and Web Scraping
requests>=2.31.0
httpx>=0.24.0
beautifulsoup4>=4.12.0
selenium>=4.10.0

# Utilities and Helpers
tqdm>=4.65.0
python-dateutil>=2.8.2
pytz>=2023.3
pathlib2>=2.3.7
jsonschema>=4.18.0

# Fingerprinting and Hashing
hashlib-compat>=1.0.1
mmh3>=4.0.0

# Optional: For better performance
orjson>=3.9.0
ujson>=5.8.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# Jupyter for analysis (optional)
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0